+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## 0.2.2 (4/6/2020)
+- Improve `AuthBloc`:
+  - Do not await AuthStorage twice in constructor
+  - Add `logOut` function
+  - Better docs
+
+## 0.2.1 (4/4/2020)
+- Added bookmarks endpoint with corresponding MockBookmarks class
+- Renamed `bookmarkStatus` to `bookmark`, `unbookmarkStatus` to `unbookmark` to closer match the Mastodon docs.
+- Moved `bookmark` and `unbookmark` methods to bookmarks endpoint
+
+## 0.2.0 (4/2/2020)
+- Major refactor for `AuthBloc`
+  - Introduce `NullInterceptorSink<T>` class to ensure that data being added to a sink cannot be null. Update `codeSink` to use the `NullInterceptorSink`.
+  - Make sure we do not call `_handleCode` until we have an authenticated application
+  - Introduce `_initialized` Completer, `initialized` Future, and `hasAccount` bool check
+  - Update `_registerApplication` with a null check/error throw
+  - Update `_handleCode` to take an `AuthenticatedApplication` as well as a code
+  - Better error handling
+  - Updated documentation
+- Removed deprecated example code. Updated code coming soon.
+
 ## 0.1.5 (3/19/2020)
 - Add missing `text` property to `Status`